Refactoring + fixing bugs + improving UI
- Added metadata(fetched from CFW) to the react-hook-form using setValue.
