+++ src/components/EditPost.tsx
@@ -2,6 +2,7 @@ import { Dispatch, MouseEvent, SetStateAction } from "react";
 import { toast } from "sonner";
 import { Pencil } from "lucide-react";
 import { signOut } from "next-auth/react";
+
 import { useVerifyToken } from "@/apiRoutes/auth-routes";
 import { Post } from "@/utils/types";
 import EditForm from "@/components/EditForm";
